Rename kromer_plot file name to sed_plot
